Add alertboc and emailjs

+++ client/src/pages/cart/Cart.tsx
@@ -9,17 +9,50 @@ import {
 } from "../../redux/actions";
 import { State } from "../../redux/store";
 import { Product } from "../../types/Product";
+import AlertBox from "../../components/alert-boxes/alertBox";
+import { useState } from "react";
+import {
+  AlertTitles,
+  AlertTypes,
+} from "../../components/alert-boxes/alertText";
+import { Form, Input, Label } from "../../style/forms";
 
 const Cart = () => {
   const cartState = useSelector((state: State) => state);
   const dispatch = useDispatch();
+  const [email, setEmail] = useState<string>("");
+  const [phoneNumber, setPhoneNumber] = useState<string>("");
+
+  const [showAlert, setShowAlert] = useState({
+    show: false,
+    type: "",
+    title: "",
+  });
 
   const handleDeleteProductFromCart = (product: Product) => {
+    setShowAlert({
+      show: true,
+      type: AlertTypes.SUCCESS,
+      title: AlertTitles.REMOVE_PROD_SUCCESS,
+    });
     dispatch(removeProductFromCart(product));
   };
 
   const handleDeleteAllProductsFromCart = () => {
-    dispatch(deleteAllProductsFromCart());
+    if (cartState.products.length > 0) {
+      setShowAlert({
+        show: true,
+        type: AlertTypes.SUCCESS,
+        title: AlertTitles.REMOVE_ALL_PRODS_FROM_CART_SUCCESS,
+      });
+      dispatch(deleteAllProductsFromCart());
+    } else {
+      setShowAlert({
+        show: true,
+        type: AlertTypes.ERROR,
+        title: AlertTitles.REMOVE_ALL_PRODS_FROM_CART_ERROR,
+      });
+    }
   };
 
   //Stripe needs http(s)
@@ -63,20 +96,54 @@ const Cart = () => {
             ))}
           </div>
           <h3>Total pris: {cartState?.total} kr</h3>
-          <div className="cart-button-container">
-            <StripeCheckout
-              token={makePayment}
-              stripeKey={process.env.REACT_APP_KEY as string}
-              name="Buy React"
-              amount={cartState.total * 100}
-            >
-              <button className="button button-primary" type="submit">
-                STRIPE
-              </button>
-            </StripeCheckout>
+          <p>
+            Du vil bli tilsendt en mail og en melding med en mappe som
+            inneholder én mp3 fil, én wav fil og én .rar fil. Vi bruker{" "}
+            <a href="https://wetransfer.com/">Wetransfer.com</a> som
+            filoverføring.
+          </p>
 
-            <button className="button button-vipps">VIPPS</button>
-          </div>
+          <Form style={{ backgroundColor: "transparent" }}>
+            <Label style={{ color: "white" }}>Epost</Label>
+            <Input
+              name="phonennumber"
+              type="text"
+              placeholder="ola-normann@gmail.com"
+              onChange={(e) => setEmail(e.target.value)}
+            />
+            <Label style={{ color: "white" }}>Mobilnummer</Label>
+            <Input
+              name="phonennumber"
+              type="text"
+              placeholder="+47 99509035"
+              onChange={(e) => setPhoneNumber(e.target.value)}
+            />
+
+            <div className="cart-button-container">
+              <StripeCheckout
+                token={makePayment}
+                stripeKey={process.env.REACT_APP_KEY as string}
+                name="Buy React"
+                amount={cartState.total * 100}
+              >
+                <button
+                  className="button button-primary"
+                  type="submit"
+                  onClick={(e) => e.preventDefault()}
+                >
+                  STRIPE
+                </button>
+              </StripeCheckout>
+
+              <button
+                className="button button-vipps"
+                onClick={(e) => e.preventDefault()}
+              >
+                VIPPS
+              </button>
+            </div>
+          </Form>
+
           <button
             className="button button-danger"
             onClick={() => handleDeleteAllProductsFromCart()}
@@ -88,6 +155,14 @@ const Cart = () => {
         <>
           <h1>Handlekurv er tom</h1>
         </>
+      )}
+
+      {showAlert.show && (
+        <AlertBox
+          type={showAlert.type}
+          title={showAlert.title}
+          onClick={() => setShowAlert({ show: false, type: "", title: "" })}
+        />
       )}
     </Wrapper>
   );
